backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
requests==2.32.3
pydantic==1.10.9
